master:

```
test keccakf_u64                     ... bench:         550 ns/iter (+/- 77) = 45 MB/s
```

this branch:

```
test keccakf_u64                     ... bench:         484 ns/iter (+/- 207) = 51 MB/s
```