@debris 2017-08-11
so optimization could be observed only on `rustc 1.21.0-nightly (1d2a6df38 2017-08-03)`. On latest nightly there is no difference between this branch and master